requirements.txt
Django==4.2.24
djangorestframework==3.15.2
django-filter==24.3
reportlab==4.2.2
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
gunicorn==22.0.0
whitenoise==6.7.0
python-dotenv==1.0.1
requests==2.32.3
<<<<<<< HEAD
=======
djangorestframework-simplejwt
>>>>>>> main
